refactor(queries): Extract response and session logic to dedicated modules
This commit refactors the questionnaires router by extracting complex database
query and mutation logic into separate query modules. The changes improve code
organization, reduce complexity in the router, and enhance modularity by:

- Moving response-related functions to `responses.ts`
- Moving session-related functions to `sessions.ts`
- Simplifying router code by using new query functions
- Maintaining existing functionality while improving code structure

The refactoring reduces duplication and makes the codebase more maintainable
by separating concerns and providing clear, focused query modules.
